Moving the creation of ACLChechker into a factory method
This is a solution that is used in both https://github.com/solid/node-solid-server/tree/feature/quota-ui and https://github.com/solid/node-solid-server/tree/bug/root-index-special-handling
