+++ lib/ldp.js
@@ -464,6 +464,14 @@ class LDP {
     return ensureNotExists(this, URI.joinPaths(containerURI, filename).toString())
   }
 
+  getTrustedOrigins (req) {
+    let trustedOrigins = [this.resourceMapper.resolveUrl(req.hostname)].concat(this.trustedOrigins)
+    if (this.multiuser) {
+      trustedOrigins.push(this.serverUri)
+    }
+    return trustedOrigins
+  }
+
   static mimeTypeIsRdf (mimeType) {
     return RDF_MIME_TYPES.has(mimeType)
   }
